+++ tests/test_protocol.py
@@ -33,6 +33,7 @@ import aioxmpp.structs as structs
 import aioxmpp.xso as xso
 import aioxmpp.nonza as nonza
 import aioxmpp.errors as errors
+import aioxmpp.utils
 
 from aioxmpp.testutils import (
     TransportMock,
@@ -297,259 +298,11 @@ class TestDebugWrapper(unittest.TestCase):
         )
 
 
-class TestAlivenessMonitor(unittest.TestCase):
-    def setUp(self):
-        self.loop = asyncio.get_event_loop()
-        self.am = protocol.AlivenessMonitor(self.loop)
-        self.listener = make_listener(self.am)
-
-    def tearDown(self):
-        del self.am
-        del self.loop
-
-    def test_defaults(self):
-        self.assertIsNone(self.am.deadtime_soft_limit)
-        self.assertIsNone(self.am.deadtime_hard_limit)
-
-    def test_soft_limit_settable(self):
-        self.am.deadtime_soft_limit = timedelta(seconds=1)
-
-        self.assertEqual(self.am.deadtime_soft_limit, timedelta(seconds=1))
-
-    def test_hard_limit_is_settable(self):
-        self.am.deadtime_hard_limit = timedelta(seconds=1)
-
-        self.assertEqual(self.am.deadtime_hard_limit, timedelta(seconds=1))
-
-    def test_soft_limit_trips(self):
-        dt = get_timeout(timedelta(seconds=0.1))
-
-        self.am.deadtime_soft_limit = dt
-        self.am.notify_received()
-
-        run_coroutine(asyncio.sleep((dt * 0.9).total_seconds()))
-
-        self.listener.on_deadtime_soft_limit_tripped.assert_not_called()
-
-        run_coroutine(asyncio.sleep((dt * 0.2).total_seconds()))
-
-        self.listener.on_deadtime_soft_limit_tripped.assert_called_once_with()
-
-    def test_notify_received_resets_soft_limit_timer(self):
-        dt = get_timeout(timedelta(seconds=0.1))
-
-        self.am.deadtime_soft_limit = dt
-        self.am.notify_received()
-
-        run_coroutine(asyncio.sleep((dt * 0.9).total_seconds()))
-
-        self.listener.on_deadtime_soft_limit_tripped.assert_not_called()
-        self.am.notify_received()
-
-        run_coroutine(asyncio.sleep((dt * 0.9).total_seconds()))
-
-        self.listener.on_deadtime_soft_limit_tripped.assert_not_called()
-
-    def test_changing_soft_limit_recaluclates_timer(self):
-        dt = get_timeout(timedelta(seconds=0.1))
-
-        self.am.deadtime_soft_limit = dt
-        self.am.notify_received()
-
-        run_coroutine(asyncio.sleep((dt * 0.9).total_seconds()))
-
-        self.listener.on_deadtime_soft_limit_tripped.assert_not_called()
-
-        self.am.deadtime_soft_limit = dt*1.5
-
-        run_coroutine(asyncio.sleep((dt * 0.5).total_seconds()))
-
-        self.listener.on_deadtime_soft_limit_tripped.assert_not_called()
-
-        run_coroutine(asyncio.sleep((dt * 0.2).total_seconds()))
-
-        self.listener.on_deadtime_soft_limit_tripped.assert_called_once_with()
-
-    def test_soft_timer_fires_even_without_any_reception(self):
-        dt = get_timeout(timedelta(seconds=0.1))
-
-        self.am.deadtime_soft_limit = dt
-
-        run_coroutine(asyncio.sleep((dt * 0.9).total_seconds()))
-
-        self.listener.on_deadtime_soft_limit_tripped.assert_not_called()
-
-        run_coroutine(asyncio.sleep((dt * 0.2).total_seconds()))
-
-        self.listener.on_deadtime_soft_limit_tripped.assert_called_once_with()
-
-    def test_hard_limit_trips(self):
-        dt = get_timeout(timedelta(seconds=0.1))
-
-        self.am.deadtime_hard_limit = dt
-        self.am.notify_received()
-
-        run_coroutine(asyncio.sleep((dt * 0.9).total_seconds()))
-
-        self.listener.on_deadtime_hard_limit_tripped.assert_not_called()
-
-        run_coroutine(asyncio.sleep((dt * 0.2).total_seconds()))
-
-        self.listener.on_deadtime_hard_limit_tripped.assert_called_once_with()
-
-    def test_notify_received_resets_hard_limit_timer(self):
-        dt = get_timeout(timedelta(seconds=0.1))
-
-        self.am.deadtime_hard_limit = dt
-        self.am.notify_received()
-
-        run_coroutine(asyncio.sleep((dt * 0.9).total_seconds()))
-
-        self.listener.on_deadtime_hard_limit_tripped.assert_not_called()
-        self.am.notify_received()
-
-        run_coroutine(asyncio.sleep((dt * 0.9).total_seconds()))
-
-        self.listener.on_deadtime_hard_limit_tripped.assert_not_called()
-
-    def test_changing_hard_limit_recaluclates_timer(self):
-        dt = get_timeout(timedelta(seconds=0.1))
-
-        self.am.deadtime_hard_limit = dt
-        self.am.notify_received()
-
-        run_coroutine(asyncio.sleep((dt * 0.9).total_seconds()))
-
-        self.listener.on_deadtime_hard_limit_tripped.assert_not_called()
-
-        self.am.deadtime_hard_limit = dt*1.5
-
-        run_coroutine(asyncio.sleep((dt * 0.5).total_seconds()))
-
-        self.listener.on_deadtime_hard_limit_tripped.assert_not_called()
-
-        run_coroutine(asyncio.sleep((dt * 0.2).total_seconds()))
-
-        self.listener.on_deadtime_hard_limit_tripped.assert_called_once_with()
-
-    def test_hard_timer_fires_even_without_any_reception(self):
-        dt = get_timeout(timedelta(seconds=0.1))
-
-        self.am.deadtime_hard_limit = dt
-
-        run_coroutine(asyncio.sleep((dt * 0.9).total_seconds()))
-
-        self.listener.on_deadtime_hard_limit_tripped.assert_not_called()
-
-        run_coroutine(asyncio.sleep((dt * 0.2).total_seconds()))
-
-        self.listener.on_deadtime_hard_limit_tripped.assert_called_once_with()
-
-    def test_both_trip(self):
-        dt = get_timeout(timedelta(seconds=0.1))
-
-        self.am.deadtime_soft_limit = dt
-        self.am.deadtime_hard_limit = dt * 1.5
-        self.am.notify_received()
-
-        run_coroutine(asyncio.sleep((dt * 0.9).total_seconds()))
-
-        self.listener.on_deadtime_soft_limit_tripped.assert_not_called()
-        self.listener.on_deadtime_hard_limit_tripped.assert_not_called()
-
-        run_coroutine(asyncio.sleep((dt * 0.2).total_seconds()))
-
-        self.listener.on_deadtime_soft_limit_tripped.assert_called_once_with()
-        self.listener.on_deadtime_hard_limit_tripped.assert_not_called()
-
-        run_coroutine(asyncio.sleep((dt * 0.4).total_seconds()))
-
-        self.listener.on_deadtime_soft_limit_tripped.assert_called_once_with()
-        self.listener.on_deadtime_hard_limit_tripped.assert_called_once_with()
-
-    def test_changing_soft_limit_may_emit_limit_immediately(self):
-        dt = get_timeout(timedelta(seconds=0.1))
-
-        self.am.notify_received()
-
-        run_coroutine(asyncio.sleep((dt*1.1).total_seconds()))
-
-        self.listener.on_deadtime_soft_limit_tripped.assert_not_called()
-        self.am.deadtime_soft_limit = dt
-
-        run_coroutine(asyncio.sleep(0))
-        self.listener.on_deadtime_soft_limit_tripped.assert_called_once_with()
-
-    def test_changing_soft_limit_does_not_reemit_hard_limit(self):
-        dt = get_timeout(timedelta(seconds=0.1))
-
-        self.am.deadtime_hard_limit = dt
-        self.am.notify_received()
-
-        run_coroutine(asyncio.sleep((dt*1.1).total_seconds()))
-
-        self.listener.on_deadtime_soft_limit_tripped.assert_not_called()
-        self.listener.on_deadtime_hard_limit_tripped.assert_called_once_with()
-
-        self.am.deadtime_soft_limit = dt * 1.5
-
-        run_coroutine(asyncio.sleep(0))
-        self.listener.on_deadtime_soft_limit_tripped.assert_not_called()
-        self.listener.on_deadtime_hard_limit_tripped.assert_called_once_with()
-
-    def test_changing_hard_limit_does_not_reemit_soft_limit(self):
-        dt = get_timeout(timedelta(seconds=0.1))
-
-        self.am.deadtime_soft_limit = dt
-        self.am.notify_received()
-
-        run_coroutine(asyncio.sleep((dt*1.1).total_seconds()))
-
-        self.listener.on_deadtime_hard_limit_tripped.assert_not_called()
-        self.listener.on_deadtime_soft_limit_tripped.assert_called_once_with()
-
-        self.am.deadtime_hard_limit = dt * 1.5
-
-        run_coroutine(asyncio.sleep(0))
-        self.listener.on_deadtime_hard_limit_tripped.assert_not_called()
-        self.listener.on_deadtime_soft_limit_tripped.assert_called_once_with()
-
-    def test_soft_limit_can_reemit_after_reception(self):
-        dt = get_timeout(timedelta(seconds=0.1))
-
-        self.am.deadtime_soft_limit = dt
-        self.am.notify_received()
-
-        run_coroutine(asyncio.sleep((dt*1.1).total_seconds()))
-
-        self.listener.on_deadtime_soft_limit_tripped.assert_called_once_with()
-        self.am.notify_received()
-        self.listener.on_deadtime_soft_limit_tripped.reset_mock()
-
-        run_coroutine(asyncio.sleep((dt*1.1).total_seconds()))
-        self.listener.on_deadtime_soft_limit_tripped.assert_called_once_with()
-
-    def test_hard_limit_can_reemit_after_reception(self):
-        dt = get_timeout(timedelta(seconds=0.1))
-
-        self.am.deadtime_hard_limit = dt
-        self.am.notify_received()
-
-        run_coroutine(asyncio.sleep((dt*1.1).total_seconds()))
-
-        self.listener.on_deadtime_hard_limit_tripped.assert_called_once_with()
-        self.am.notify_received()
-        self.listener.on_deadtime_hard_limit_tripped.reset_mock()
-
-        run_coroutine(asyncio.sleep((dt*1.1).total_seconds()))
-        self.listener.on_deadtime_hard_limit_tripped.assert_called_once_with()
-
-
 class TestXMLStream(unittest.TestCase):
     def setUp(self):
         self.maxDiff = None
         self.loop = asyncio.get_event_loop()
-        self.monitor = unittest.mock.Mock(spec=protocol.AlivenessMonitor)
+        self.monitor = unittest.mock.Mock(spec=aioxmpp.utils.AlivenessMonitor)
 
     def tearDown(self):
         del self.monitor
@@ -584,7 +337,7 @@ class TestXMLStream(unittest.TestCase):
         with contextlib.ExitStack() as stack:
             if monitor_mock:
                 AlivenessMonitor = stack.enter_context(
-                    unittest.mock.patch("aioxmpp.protocol.AlivenessMonitor")
+                    unittest.mock.patch("aioxmpp.utils.AlivenessMonitor")
                 )
                 AlivenessMonitor.return_value = self.monitor
             else:
